Relocated temporary file class to INTV.TestHelpers.Shared
