+++ Tests/INTV.TestHelpers.Shared/INTV.TestHelpers.Shared.csproj
@@ -41,6 +41,7 @@
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Utility\ICompressedArchiveTest.cs" />
+    <Compile Include="Utility\TemporaryFile.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\INTV.Shared\INTV.Shared.desktop.csproj">
